Move constant to `virtool.example`
* Remove `virtool.example.paths`
