+++ tests/fixtures/core.py
@@ -7,7 +7,7 @@ import multidict
 import functools
 from concurrent.futures import ThreadPoolExecutor
 
-from virtool.example.paths import example_path as path
+from virtool.example import example_path as path
 
 SAM_PATH = Path.cwd() / "tests" / "test_files" / "test_al.sam"
 SAM_50_PATH = Path.cwd() / "tests" / "test_files" / "sam_50.sam"
