+++ virtool/example.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+example_path = Path(__file__).parent.parent / "example"
