--- virtool/example/paths.py
@@ -1,3 +1,0 @@
-from pathlib import Path
-
-example_path = Path(__file__).parent.parent.parent / "example"
